+++ tests/testthat/test-valid-term.R
@@ -4,7 +4,7 @@ test_that("valid_term", {
   expect_error(valid_term(NA_character_), "^`x` must inherit from S3 class 'term'[.]$",
     class = "chk_error"
   )
-  expect_identical(valid_term(term(0L)), TRUE)
+  expect_identical(valid_term(term(0L)), logical(0))
   expect_identical(valid_term(as.term(c("a", NA_character_))), c(TRUE, NA))
   expect_identical(
     valid_term(as.term(c("a", "a [3]", " b [ 1  ] ", "c[1]"))),