+++ NEWS.md
@@ -1,3 +1,4 @@
+- Added `scalar_term()` to test if each term element is scalar.
 - Added `pars_terms()`.
 - Migrated `ndims()` to extras package.
 - Migrated `dims()` to extras package.